Now saves inSPIRE pages that you pass with -a
Goal: you pass it an inSPIRE page with the -a option, and it fetches the bibtex for that paper and adds it to the bibtex file.
The bibtex file is opened AFTER this, so the abstract and pdf will be downloaded as usual.

FOR NOW it only saves the inspire page, but doesn't retrieve data from it.