Change: now gets bibtex without saving webpages
Retreives from the pages 'live', without saving.
First gets accesses the link given and finds the link to the page with the bibtex
Second accesses the bibtex page and extracts the pure bibtex text.

BUT what should be done with it then?
It's most elegant to have the URLs processed one by one by readoptions, but the bibtex file isn't made ready until later on in main.
But the bibtex wants to go in the bibtex file, obviously.
So how should I order things?