Restructured to store URLs in "LISTOFINSPIREURLS"
This gets around the problem of ordering of commands in a relatively elegant way.
Had to move around a lot of the echo commands.
Added random waits of around 5 seconds to the inSPIRE wget commands
TO DO: make it check if the entry is already in the bibfile before appending it to the bibfile. Can use the UID.